Refactor BingoBoardUI for singificantly improved clarity
